+++ SandboxiePlus/SandMan/SbiePlusAPI.h
@@ -145,6 +145,7 @@ public:
 
 	EBoxTypes				GetType() const { return m_BoxType; }
 	bool					IsAutoDelete() const { return m_BoxDel; }
+	bool					IsForceDisabled() const { return m_NoForce; }
 	QRgb					GetColor() const { return m_BoxColor; }
 	
 	class COptionsWindow*	m_pOptionsWnd;
@@ -229,5 +230,6 @@ protected:
 
 	EBoxTypes				m_BoxType;
 	bool					m_BoxDel;
+	bool					m_NoForce;
 	QRgb					m_BoxColor;
 };